Fix: prevent infinite loop
If a category is both a parent and a child of another category
the program falls into an infinite loop. To avoid this, new categories
data must be cleaned by hand up before using them.
Now the program automatically discards a category as sub category if it is
also a parent category.